Changed .Rdw extension to .rnw
I'd like this project to be recognized as an R project
on GitHub, but all of the source files were in a
nonstandard extension.  Hopefully, changing the extensions
to .rnw will fix the language indexing problem... unfortunately,
these extensions are somewhat misleading since these are
noweb files and not Sweave files.  But I think the tradeoff
is potentially worth it, and I can always change the
extension again.